test: relocate helpers
