+++ test/unit/__helper/index.js
@@ -1,5 +1,5 @@
 import path from 'path';
 
 export const parseConfigurationFile = (name) => {
-  return path.join(__dirname, `./partials/config/yaml/${name}.yaml`);
+  return path.join(__dirname, `../partials/config/yaml/${name}.yaml`);
 };
