+++ test/unit/__helper/mock.js
@@ -9,9 +9,11 @@ import type {IServerBridge} from '../../types';
 
 export function mockServer(port: number) {
   const pathStore = path.join(__dirname, '../partials');
-  const verdaccioConfig = new VerdaccioConfig(
-    path.join(pathStore, '/mock-store'),
-    path.join(pathStore, '/config-unit-mock-server-test.yaml'), `http://${DOMAIN_SERVERS}:${port}/`, port);
+  const storePath = path.join(pathStore, '/mock-store');
+  const configPath = path.join(pathStore, '/config-unit-mock-server-test.yaml');
+
+  const verdaccioConfig = new VerdaccioConfig(storePath, configPath, `http://${DOMAIN_SERVERS}:${port}/`, port);
+
   const server: IServerBridge = new Server(verdaccioConfig.domainPath);
 
   return new VerdaccioProcess(verdaccioConfig, server, false, false, false);
