+++ test/unit/api/api.spec.js
@@ -9,10 +9,10 @@ import starMetadata from '../partials/star-api';
 import endPointAPI from '../../../src/api/index';
 
 import {HEADERS, API_ERROR, HTTP_STATUS, HEADER_TYPE, API_MESSAGE, TOKEN_BEARER} from '../../../src/lib/constants';
-import {mockServer} from './mock';
+import {mockServer} from '../__helper/mock';
 import {DOMAIN_SERVERS} from '../../functional/config.functional';
 import {buildToken} from '../../../src/lib/utils';
-import {getNewToken} from './__api-helper';
+import {getNewToken} from '../__helper/api';
 
 require('../../../src/lib/logger').setup([]);
 const credentials = { name: 'jota', password: 'secretPass' };
