+++ test/unit/api/up-storage.spec.js
@@ -9,7 +9,7 @@ import {setup} from '../../../src/lib/logger';
 import type {Config, UpLinkConf} from '@verdaccio/types';
 import type {IProxy} from '../../../types/index';
 import {API_ERROR, HTTP_STATUS} from "../../../src/lib/constants";
-import {mockServer} from './mock';
+import {mockServer} from '../__helper/mock';
 import {DOMAIN_SERVERS} from '../../functional/config.functional';
 
 setup([]);
