+++ test/unit/functionalLibs/request.spec.js
@@ -2,7 +2,7 @@
 
 import _ from 'lodash';
 import smartRequest, {PromiseAssert} from '../../lib/request';
-import {mockServer} from '../api/mock';
+import {mockServer} from '../__helper/mock';
 import {HTTP_STATUS} from '../../../src/lib/constants';
 import type {IRequestPromise} from '../../types';
